Fixes #10 a workaround that fixes the immediate svg rendering problem with a hidden div
My solution is not optimal, but at least fixes the immediate rendering
problem. The difference between "display: none" and "visibility: hidden"
is discussed here: http://svg-whiz.com/svg/HideShow.svg

The reason it is not optimal is because visibility still uses the same
amount of space whether hidden or not so there is a large blank space
where the hidden material will appear.